web/secure-serializer/app/requirements.txt
flask
python-dotenv
uwsgi
